+++ react/ResultsList.js
@@ -1,0 +1,3 @@
+import ResultsList from './components/ResultsList/index'
+
+export default ResultsList