+++ react/components/SearchBar/components/SearchBar.js
@@ -3,16 +3,24 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import Downshift from 'downshift'
 import debounce from 'debounce'
-import { NoSSR } from 'vtex.render-runtime'
+import {
+  NoSSR,
+  useRuntime,
+  ExtensionPoint,
+  useChildBlock,
+} from 'vtex.render-runtime'
 import { Overlay } from 'vtex.react-portal'
-import { useRuntime } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
+import { intlShape } from 'react-intl'
+
+import ResultsList from '../../ResultsList'
 
 import AutocompleteInput from './AutocompleteInput'
-import ResultsLists from './ResultsList'
 
 const CSS_HANDLES = ['searchBarContainer', 'searchBarInnerContainer']
 const SEARCH_DELAY_TIME = 500
+const AUTCOMPLETE_EXTENSION_ID = 'autocomplete-result-list'
+const INPUT_ERROR_MESSAGE_DELAY = 2000
 
 const SearchBar = ({
   placeholder,
@@ -28,11 +36,20 @@ const SearchBar = ({
   maxWidth,
   attemptPageTypeSearch,
   customSearchPageUrl,
+  autocompleteAlignment,
+  openAutocompleteOnFocus,
+  blurOnSubmit,
+  submitOnIconClick,
+  minSearchTermLength,
+  autocompleteFullWindow,
+  intl,
 }) => {
   const container = useRef()
   const { navigate } = useRuntime()
   const handles = useCssHandles(CSS_HANDLES)
   const [searchTerm, setSearchTerm] = useState(inputValue)
+  const [inputErrorMessage, setInputErrorMessage] = useState()
+  const inputErrorMessageTimeRef = useRef()
 
   const debouncedSetSearchTerm = useCallback(
     debounce(newValue => {
@@ -97,6 +114,36 @@ const SearchBar = ({
     [navigate, attemptPageTypeSearch, customSearchPageUrl]
   )
 
+  const validateInput = () => {
+    if (minSearchTermLength && inputValue.length < minSearchTermLength) {
+      return intl.formatMessage({
+        id: 'store/search.searchTermTooShort',
+      })
+    }
+
+    return undefined
+  }
+
+  const showInputErrorMessageTemporarily = inputErrorMessage => {
+    setInputErrorMessage(inputErrorMessage)
+
+    if (inputErrorMessageTimeRef.current) {
+      clearTimeout(inputErrorMessageTimeRef.current)
+    }
+
+    inputErrorMessageTimeRef.current = setTimeout(() => {
+      setInputErrorMessage()
+    }, INPUT_ERROR_MESSAGE_DELAY)
+  }
+
+  const hideInputErrorMessage = () => {
+    if (inputErrorMessageTimeRef.current) {
+      clearTimeout(inputErrorMessageTimeRef.current)
+    }
+
+    setInputErrorMessage()
+  }
+
   const fallback = (
     <AutocompleteInput
       placeholder={placeholder}
@@ -105,8 +152,13 @@ const SearchBar = ({
       hasIconLeft={hasIconLeft}
       iconClasses={iconClasses}
       iconBlockClass={iconBlockClass}
+      inputErrorMessage={inputErrorMessage}
     />
   )
+
+  const SelectedResultsList = useChildBlock({ id: AUTCOMPLETE_EXTENSION_ID })
+    ? props => <ExtensionPoint id={AUTCOMPLETE_EXTENSION_ID} {...props} />
+    : props => <ResultsList {...props} />
 
   return (
     <div
@@ -128,6 +180,7 @@ const SearchBar = ({
             highlightedIndex,
             isOpen,
             closeMenu,
+            openMenu,
           }) => (
             <div
               className={classNames(
@@ -142,22 +195,44 @@ const SearchBar = ({
                 onClearInput={onClearInput}
                 hasIconLeft={hasIconLeft}
                 iconClasses={iconClasses}
+                onGoToSearchPage={onGoToSearchPage}
+                submitOnIconClick={submitOnIconClick}
+                openAutocompleteOnFocus={openAutocompleteOnFocus}
+                openMenu={openMenu}
+                inputErrorMessage={inputErrorMessage}
                 {...getInputProps({
                   onKeyDown: event => {
                     // Only call default search function if user doesn't
                     // have any item highlighted in the menu options
                     if (event.key === 'Enter' && highlightedIndex === null) {
+                      const errorMessage = validateInput()
+
+                      if (errorMessage) {
+                        showInputErrorMessageTemporarily(errorMessage)
+                        return
+                      }
+
+                      if (blurOnSubmit) {
+                        event.currentTarget.blur()
+                      }
+
                       onGoToSearchPage()
                       closeMenu()
+                    } else {
+                      hideInputErrorMessage()
                     }
                   },
                   placeholder,
                   value: inputValue,
                   onChange: onInputChange,
+                  onFocus: openAutocompleteOnFocus ? openMenu : undefined,
                 })}
               />
-              <Overlay alignment="right">
-                <ResultsLists
+              <Overlay
+                alignment={autocompleteAlignment}
+                fullWindow={autocompleteFullWindow}
+              >
+                <SelectedResultsList
                   parentContainer={container}
                   {...{
                     attemptPageTypeSearch,
@@ -206,8 +281,24 @@ SearchBar.propTypes = {
   maxWidth: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   /** A template for a custom url. It can have a substring ${term} used as placeholder to interpolate the searched term. (e.g. `/search?query=${term}`) */
   customSearchPageUrl: PropTypes.string,
-  iconBlockClass: PropTypes.string,
+  /** Uses the term the user has inputed to try to navigate to the proper
+   * page type (e.g. a department, a brand, a category)
+   */
   attemptPageTypeSearch: PropTypes.bool,
+  /* Autocomplete Horizontal alignment */
+  autocompleteAlignment: PropTypes.string,
+  /** Identify if autocomplete should be open on input focus or not */
+  openAutocompleteOnFocus: PropTypes.bool,
+  /** Identify if input should blur on submit */
+  blurOnSubmit: PropTypes.bool,
+  /** Identify if icon should submit on click */
+  submitOnIconClick: PropTypes.bool,
+  /** Minimum search term length allowed */
+  minSearchTermLength: PropTypes.number,
+  /** If true, the autocomplete will fill the whole window horizontally */
+  autocompleteFullWindow: PropTypes.bool,
+  /* Internationalization */
+  intl: intlShape.isRequired,
 }
 
 export default SearchBar