+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `minSearchTermLength` prop to `SearchBar`. Defines the minimum search term length allowed.
+- `autocompleteAlignment` prop to `SearchBar`. Autocomplete Horizontal alignment.
+- `openAutocompleteOnFocus` prop to `SearchBar`. Identify if autocomplete should be open on input focus or not.
+- `blurOnSubmit` prop to `SearchBar`. Identify if input should blur on submit.
+- `autocompleteFullWidth` prop to `SearchBar`. If true, the autocomplete will fill the whole window width.
+- `submitOnIconClick` prop to `SearchBar`. Identify if icon should submit on click.
+
+### Changed
+- Now, the search bar input and the autocomplete are different components. `search-bar` and `autocomplete-result-list` respectively.
 
 ## [3.96.2] - 2019-12-20
 ### Changed